requirements.txt
aenum>=2.1.2
flaky>=3.4.0
bunch>=1.0.1
psutil>=3.4.2
parameterized>=0.6.1
nose>=1.3.7
nose-html-reporting>=0.2.3
Pillow>=4.0.0
numpy>=1.11.1
opencv-python>=3.4.3.18
pytesseract>=0.1.6